pyproject.toml
[project]
name = "paper-review"
version = "0.1.0"
description = "Personal paper review system (Drive + DOI metadata + single-session multi-persona analysis)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "typer>=0.12",
  "rich>=13.7",
  "itsdangerous>=2.2",
  "google-auth>=2.34",
  "google-auth-oauthlib>=1.2",
]

[project.scripts]
paper-review = "paper_review.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
paper_review = ["webapp/**/*.html", "webapp/**/*.css", "webapp/**/*.js"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
